mrc-spec-aux-hidden-state(extract-spec-aux-hidden-state-move,mechanical_provable): move spec aux-hidden-state resolution to spec_aux_hidden_state module (cut+paste)
